app/requirements.txt
# CervixAI Dependencies
# Core Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.28.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Frontend
streamlit>=1.28.0
requests>=2.31.0

# Image Processing
pillow>=10.0.0
numpy>=1.24.0

# PDF Reports
reportlab>=4.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0
